refactor(frontend): move agent access cache hooks
